move errorchecker and staticlint to pkg
