+++ cmd/staticlint/main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"github.com/shulganew/shear.git/pkg/staticlint"
+	"golang.org/x/tools/go/analysis/multichecker"
+)
+
+func main() {
+	multichecker.Main(staticlint.NewAnalyzer()...)
+}
