+++ pkg/errcheckanalyzer/errcheckanalyzer_test.go
@@ -1,4 +1,4 @@
-package main
+package errcheckanalyzer
 
 import (
 	"testing"
