+++ pkg/errcheckanalyzer/testdata/pkg1/pkg1.go
