+++ pkg/staticlint/staticlint.go
@@ -1,14 +1,13 @@
-package main
+package staticlint
 
 import (
-	"flag"
 	"fmt"
-	"go/ast"
 	"strings"
 
 	"4d63.com/gochecknoglobals/checknoglobals"
 	"github.com/butuzov/ireturn/analyzer"
 	"github.com/kyoh86/exportloopref"
+	"github.com/shulganew/shear.git/pkg/errcheckanalyzer"
 	"golang.org/x/tools/go/analysis"
 	"golang.org/x/tools/go/analysis/multichecker"
 	"golang.org/x/tools/go/analysis/passes/appends"
@@ -62,14 +61,7 @@ import (
 	"honnef.co/go/tools/staticcheck"
 )
 
-//go:generate go run ./... -d
-func main() {
-
-	dgen := flag.Bool("d", false, "flag for doc genetation")
-	flag.Parse()
-
-	var analyzers []*analysis.Analyzer
-
+func NewAnalyzer() (analyzers []*analysis.Analyzer) {
 	// Analysis package, passes group.
 	passes := []*analysis.Analyzer{appends.Analyzer, asmdecl.Analyzer, httpmux.Analyzer, assign.Analyzer, atomic.Analyzer, atomicalign.Analyzer, bools.Analyzer, buildssa.Analyzer, buildtag.Analyzer, cgocall.Analyzer, composite.Analyzer, copylock.Analyzer, ctrlflow.Analyzer, deepequalerrors.Analyzer, defers.Analyzer, directive.Analyzer, errorsas.Analyzer, fieldalignment.Analyzer, findcall.Analyzer, framepointer.Analyzer, httpresponse.Analyzer, ifaceassert.Analyzer, inspect.Analyzer, loopclosure.Analyzer, lostcancel.Analyzer, nilfunc.Analyzer, nilness.Analyzer, pkgfact.Analyzer, printf.Analyzer, reflectvaluecompare.Analyzer, shadow.Analyzer, shift.Analyzer, sigchanyzer.Analyzer, slog.Analyzer, sortslice.Analyzer, stdmethods.Analyzer, stringintconv.Analyzer, structtag.Analyzer, testinggoroutine.Analyzer, tests.Analyzer, timeformat.Analyzer, unmarshal.Analyzer, unreachable.Analyzer, unsafeptr.Analyzer, unusedresult.Analyzer, unusedwrite.Analyzer, usesgenerics.Analyzer}
 
@@ -98,13 +90,6 @@ func main() {
 	// Public Analizers.
 	var pub []*analysis.Analyzer
 
-	// Custom analizer.
-	var osExitCheckAnalyzer = &analysis.Analyzer{
-		Name: "osExit",
-		Doc:  "Check if os.Exit call exist in the main package for the main function.",
-		Run:  run,
-	}
-
 	pub = append(pub, exportloopref.Analyzer)
 	pub = append(pub, checknoglobals.Analyzer())
 	pub = append(pub, analyzer.NewAnalyzer())
@@ -113,62 +98,16 @@ func main() {
 	analyzers = append(analyzers, analyserSA...)
 	analyzers = append(analyzers, quick...)
 	analyzers = append(analyzers, pub...)
+
+	// Adding custom analyser osExit().
+	var osExitCheckAnalyzer = errcheckanalyzer.NewAnalyzer()
 	analyzers = append(analyzers, osExitCheckAnalyzer)
 
-	if !*dgen {
-		// Create custom multichecker.
-		fmt.Println("Analizers from go/analysis/passes: ", len(passes))
-		fmt.Println("Staticcheck analyzers SA*: ", len(analyserSA))
-		fmt.Println("Analizers quick checks: ", len(quick))
-		fmt.Println("Analizers public checks: ", len(pub))
-		multichecker.Main(analyzers...)
-	} else {
-		// Generate doc.go.
-		fmt.Println("Generate user's documentation: ")
-		docgen(analyzers...)
-	}
-
-}
-
-func run(pass *analysis.Pass) (interface{}, error) {
-	for _, file := range pass.Files {
-		filename := pass.Fset.Position(file.Pos()).Filename
-		if !strings.HasSuffix(filename, ".go") {
-			continue
-		}
-		// Finding main package.
-		ast.Inspect(file, func(node ast.Node) bool {
-			if f, ok := node.(*ast.File); ok {
-				if f.Name.Name == "main" {
-					// Finding main method.
-					ast.Inspect(node, func(pmnode ast.Node) bool {
-						if fd, ok := pmnode.(*ast.FuncDecl); ok {
-							if fd.Name.Name == "main" {
-								// Inspect main method.
-								ast.Inspect(pmnode, func(mnode ast.Node) bool {
-									// if os.Exit in go statement
-									if _, ok := mnode.(*ast.GoStmt); ok {
-										return false
-									}
-									if sl, ok := mnode.(*ast.SelectorExpr); ok {
-										if ind, ok := sl.X.(*ast.Ident); ok {
-											if ind.Name == "os" && sl.Sel.Name == "Exit" {
-												pass.Reportf(sl.Pos(), "strait call os.Exit in the main method prohibited")
-											}
-										}
-									}
-									return true
-								})
-								return false
-							}
-						}
-						return true
-					})
-					return false
-				}
-			}
-			return true
-		})
-	}
-	return nil, nil
+	// Create custom multichecker.
+	fmt.Println("Analizers from go/analysis/passes: ", len(passes))
+	fmt.Println("Staticcheck analyzers SA*: ", len(analyserSA))
+	fmt.Println("Analizers quick checks: ", len(quick))
+	fmt.Println("Analizers public checks: ", len(pub))
+	multichecker.Main(analyzers...)
+	return
 }
